VolumeScan, Spiral scan and downsampling
- SpiralScan tested  on 200 khz system with FPGA processing - it
completes with no errors but unable to verify that image is correct

- Added downsampling support in OCTFPGAProcessing and implemented in
Bscan and VolumeScan

- Tested volume rotation - it appears to work

- Volume processing now holds both uncorrected and correct aspect ratio.
volumes.  The corrected aspect ratio changes the Z instead of X
(different than BScan) .  For a square X/Y image, this is very fast and
typically generates samller images which is helpful for the 3D viewer as
it tends to crash if handed too larger a volume.    This may need to be
reworked with a full 3D interpolation after entire volume has been
collected.

- Changing the 3D norms will now update volume viewer properly