Long acquisition issue fixed / VolumeScan
- Fixed issue when acquiring data over a long period of time, that the
program would freeze for several seconds/minutes.   This was caused  due
to the data buffer being reallocated for every AcquireXXX call in
LV_DLLInterface.    This most likely had to do something with clearing
the heap after the runXXX function exited, and the local variables went
out of scope.  The Python interpreter would have to free up unused
blocks on the heap.    Now there is an internal state variable in
LV_DLLInterface, which will only reallocate the buffere if the previous
acquisition was a different type or used different acquisition
parameters that affected the size of the buffer.   There still appears
to be a big of lag (a few seconds) when stopping a long acquisition, but
nothing like before.

- VolumeScan multiprocess fixed various issues related to stopping and
restarting the acquisition,  caused by sending the processing message
queue with updated  procOpts every loop iteration.   Now the procOpts is
only sent if  it is different than the previou cycle.