BScan software collection
Fixed bug in BScan where the example data was being loaded even if when
not in OCT testing mode.   This caused the software BScan to fail on the
200 kHz system, since loading the data was also setting the requested
samples per trig..  Since the example data was saved on the 50 kHz
system, it set a value for requested samples per trigger much higher
than what the 200 kHz FPGA code could handle (was set to ~2100 when 200
kHz should be about 1300 samples/trig) .  As a result, junk data was
returned.